Readme
Restaurants frequently face the problem of either wastage of food on one hand or devoiding their
customers to provide enough choice because some items go out of stock. Sometimes, the
customers have to wait a lot on busy days due to lack of staff and resources to handle traffic. The
problem of the restaurant business is the lack of effective prediction about the number of customers they
would have in advance. In order to help restaurants to keep better management in their place, we
try to predict the approximate range of visitors will be visiting the restaurant for future dates using
multiple datasets from different sources